testcases performed

+++ src/main/java/com/stackroute/muzixservice/service/MusicServiceImpl.java
@@ -1,6 +1,7 @@
 package com.stackroute.muzixservice.service;
 
 import com.stackroute.muzixservice.domain.Music;
+import com.stackroute.muzixservice.exceptions.TrackAlreadyExistsException;
 import com.stackroute.muzixservice.exceptions.TrackNotFoundException;
 import com.stackroute.muzixservice.repository.MusicRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,6 +13,7 @@ import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
 import javax.annotation.PostConstruct;
 import java.util.List;
+import java.util.Optional;
 
 @CacheConfig(cacheNames = "music")
 @Component
@@ -35,42 +37,52 @@ public class MusicServiceImpl implements MusicService
     }
 
     @Override
-    public Music saveTrack(Music music)
-    {
-        Music savedTrack = musicRepository.save(music);
+    public Music saveTrack(Music music)throws TrackAlreadyExistsException {
+        if(musicRepository.existsById(music.getId()))
+        {
+            throw new TrackAlreadyExistsException("User Already Exists");
+        }
+        Music savedTrack=musicRepository.save(music);
+        if(savedTrack==null)
+        {
+            throw new TrackAlreadyExistsException("User Already Exists");
+        }
         return savedTrack;
     }
 
     @Override
     @Cacheable
-    public List<Music> displayTrack() {
+    public List<Music> displayTrack() throws TrackNotFoundException{
         return musicRepository.findAll();
     }
 
-//    @Override
-//    public Music displayTracksById(int id) {
-//        return musicRepository.findById(id).get();
-//    }
-
     @Override
     @CachePut
-    public Music updateTrackComments(Music music, int id)throws TrackNotFoundException {
-        if(!musicRepository.existsById(music.getId()))
+    public Music updateTrackComments(Music music, int id) throws TrackNotFoundException {
+        Music musicList=null;
+        Optional optional=musicRepository.findById(id);
+        if(optional.isPresent())
         {
-            throw new TrackNotFoundException("User Not Found");
+            musicList = musicRepository.findById(id).get();
+            musicList.setTrackComments(music.getTrackComments());
         }
-        Music musicList = musicRepository.findById(id).get();
-        musicList.setTrackComments(music.getTrackComments());
         return musicList;
     }
 
     @Override
-    public void removeTrack(int trackId) {
-        musicRepository.deleteById(trackId);
+    public Music removeTrack(int trackId)throws TrackNotFoundException{
+        Music music=null;
+        Optional optional=musicRepository.findById(trackId);
+        if(optional.isPresent())
+        {
+            music = musicRepository.findById(trackId).get();
+            musicRepository.deleteById(trackId);
+        }
+        return music;
     }
 
     @Override
-    public Music trackByName(String trackName) {
+    public Music trackByName(String trackName) throws TrackNotFoundException{
         return musicRepository.trackByName(trackName);
     }
